associa login e às classes que são usuarios do sistema

+++ src/main/java/com/ifsudestemg/ecommerce/example/ecommerceapi/model/repository/LoginRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface LoginRepository  extends JpaRepository<Login, Integer> {
+public interface LoginRepository  extends JpaRepository<Login, Long> {
     Optional<Login> findByLogin(String login);
 }
